requirements.txt
torch
torchvision
pyyaml
numpy
pillow
scikit-learn
matplotlib
